requirements.txt
pbr>=1.6.0
twine

oslo.config>=1.9.0
lockfile
six>=1.9.0
stevedore>=1.5.0
titlecase>=0.7.2
tinydb<3.0,>=2.4
PrettyTable<0.8,>=0.7

trakt.py>=2.5.1
tvdbapi_client>=0.1.4

